requirements.txt
blinker==1.4
cffi==1.14.2
click==7.1.2
cryptography==3.1
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Ext==0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
Pillow==7.2.0
pycountry==20.7.3
pycparser==2.20
pymongo==3.11.0
qrcode==6.1
sentinels==1.0.0
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.3
